Fix PR feedback - Push to extentions and rename method call
